move tests, fix fmt
